Note: last few commits had a hickup
Changed the defaultBoxes function to have a nested for loop. The first
creating a set of cells and the second one looping a set of boxes, which are then appended
to the cells of the first loop.

I also, changed the querySelectorAll, in deleteBoxes(), coloredBoxes(), resetButton()
from .new to .actualBox