Adjust resetButton() so that random color gets cleared
Prior to using the randomColor(), which does inline styling, I was using the CSS file
to change the styling (background Color) of a div. I used to remove with the clear button
the class that added the styling rules upon mouseenter. Now it aditionally, sets the background
color to it's default upon pressing the clear button.